refactor(frontend): drop the views/cluster folder
Drop the views/cluster folder and move all children up to views/
