+++ frontend/src/pages/(authenticated)/clusters/[cluster]/machine/[machine]/monitor.vue
@@ -21,7 +21,7 @@ import { MachineService, type ProcessInfo } from '@/api/talos/machine/machine.pb
 import PageContainer from '@/components/PageContainer/PageContainer.vue'
 import { getContext } from '@/context'
 import { formatBytes } from '@/methods'
-import NodesMonitorChart from '@/views/cluster/Nodes/components/NodesMonitorChart.vue'
+import NodesMonitorChart from '@/views/Nodes/components/NodesMonitorChart.vue'
 
 definePage({ name: 'NodeMonitor' })
