+++ frontend/src/pages/(authenticated)/clusters/[cluster]/patches.vue
@@ -7,7 +7,7 @@ included in the LICENSE file.
 <script setup lang="ts">
 import PageContainer from '@/components/PageContainer/PageContainer.vue'
 import PageHeader from '@/components/PageHeader.vue'
-import Patches from '@/views/cluster/Config/Patches.vue'
+import Patches from '@/views/Config/Patches.vue'
 
 definePage({ name: 'ClusterConfigPatches' })
 </script>
