+++ frontend/src/pages/(authenticated)/clusters/[cluster]/scale.vue
@@ -25,6 +25,7 @@ import {
   MachineStatusType,
 } from '@/api/resources'
 import { itemID } from '@/api/watch'
+import ManagedByTemplatesWarning from '@/components/ManagedByTemplatesWarning.vue'
 import PageContainer from '@/components/PageContainer/PageContainer.vue'
 import PageHeader from '@/components/PageHeader.vue'
 import TSpinner from '@/components/Spinner/TSpinner.vue'
@@ -33,7 +34,6 @@ import Watch from '@/components/Watch/Watch.vue'
 import { ClusterCommandError, clusterSync } from '@/methods/cluster'
 import { showError, showSuccess } from '@/notification'
 import { populateExisting, state } from '@/states/cluster-management'
-import ManagedByTemplatesWarning from '@/views/cluster/ManagedByTemplatesWarning.vue'
 import ClusterMenu from '@/views/Clusters/ClusterMenu.vue'
 import ClusterMachineItem from '@/views/Clusters/Management/ClusterMachineItem.vue'
 import MachineSets from '@/views/Clusters/Management/MachineSets.vue'
