+++ frontend/src/pages/(authenticated)/machines/[machine]/patches/[patch].vue
@@ -5,7 +5,7 @@ Use of this software is governed by the Business Source License
 included in the LICENSE file.
 -->
 <script setup lang="ts">
-import PatchEdit from '@/views/cluster/Config/PatchEdit.vue'
+import PatchEdit from '@/views/Config/PatchEdit.vue'
 
 definePage({ name: 'MachinePatchEdit' })
 </script>
