+++ frontend/src/views/Clusters/ClusterItem.vue
@@ -22,7 +22,7 @@ import { downloadKubeconfig, downloadTalosconfig } from '@/methods'
 import { useClusterPermissions } from '@/methods/auth'
 import { addClusterLabels, removeClusterLabels } from '@/methods/cluster'
 import type { Label } from '@/methods/labels'
-import ClusterMachines from '@/views/cluster/ClusterMachines/ClusterMachines.vue'
+import ClusterMachines from '@/views/ClusterMachines/ClusterMachines.vue'
 import ClusterStatus from '@/views/Clusters/ClusterStatus.vue'
 import ItemLabels from '@/views/ItemLabels/ItemLabels.vue'
 import ClusterDestroy from '@/views/Modals/ClusterDestroy.vue'
