+++ frontend/src/views/Modals/MaintenanceUpdate.vue
@@ -16,11 +16,11 @@ import type { MachineStatusSpec, TalosVersionSpec } from '@/api/omni/specs/omni.
 import { DefaultNamespace, MachineStatusType, TalosVersionType } from '@/api/resources'
 import TButton from '@/components/Button/TButton.vue'
 import TCheckbox from '@/components/Checkbox/TCheckbox.vue'
+import ManagedByTemplatesWarning from '@/components/ManagedByTemplatesWarning.vue'
 import TSpinner from '@/components/Spinner/TSpinner.vue'
 import { majorMinorVersion } from '@/methods'
 import { useResourceWatch } from '@/methods/useResourceWatch'
 import { showError, showSuccess } from '@/notification'
-import ManagedByTemplatesWarning from '@/views/cluster/ManagedByTemplatesWarning.vue'
 import CloseButton from '@/views/Modals/CloseButton.vue'
 
 const router = useRouter()
