+++ frontend/src/views/Nodes/NodesHeader.vue
@@ -130,7 +130,7 @@ const { canRebootMachines, canRemoveMachines, canAddClusterMachines } = useClust
 </template>
 
 <style scoped>
-@reference "../../../index.css";
+@reference "../../index.css";
 
 .header-button:first-child {
   border-radius: 4px 0 0 4px;
