+++ frontend/src/views/Nodes/components/DiskPartitionTable.vue
