+++ frontend/src/views/Nodes/components/DiskUsageBar.vue
