+++ frontend/src/views/Nodes/components/NodeConditions.vue
