+++ frontend/src/views/Overview/components/OverviewRightPanel/OverviewRightPanel.vue
@@ -33,6 +33,7 @@ import {
 } from '@/api/resources'
 import TButton from '@/components/Button/TButton.vue'
 import TIcon from '@/components/Icon/TIcon.vue'
+import ManagedByTemplatesWarning from '@/components/ManagedByTemplatesWarning.vue'
 import TSpinner from '@/components/Spinner/TSpinner.vue'
 import Tooltip from '@/components/Tooltip/Tooltip.vue'
 import type { BackupsStatus } from '@/methods'
@@ -43,13 +44,12 @@ import { controlPlaneMachineSetId } from '@/methods/machineset'
 import { formatISO } from '@/methods/time'
 import { useResourceWatch } from '@/methods/useResourceWatch'
 import { showError, showWarning } from '@/notification'
-import ManagedByTemplatesWarning from '@/views/cluster/ManagedByTemplatesWarning.vue'
-import DownloadSupportBundleModal from '@/views/cluster/Overview/components/OverviewRightPanel/DownloadSupportBundleModal.vue'
-import OverviewOIDCToast from '@/views/cluster/Overview/components/OverviewRightPanel/OverviewOIDCToast.vue'
-import OverviewRightPanelCondition from '@/views/cluster/Overview/components/OverviewRightPanel/OverviewRightPanelCondition.vue'
-import OverviewRightPanelItem from '@/views/cluster/Overview/components/OverviewRightPanel/OverviewRightPanelItem.vue'
 import TClusterStatus from '@/views/Clusters/ClusterStatus.vue'
 import ClusterDestroy from '@/views/Modals/ClusterDestroy.vue'
+import DownloadSupportBundleModal from '@/views/Overview/components/OverviewRightPanel/DownloadSupportBundleModal.vue'
+import OverviewOIDCToast from '@/views/Overview/components/OverviewRightPanel/OverviewOIDCToast.vue'
+import OverviewRightPanelCondition from '@/views/Overview/components/OverviewRightPanel/OverviewRightPanelCondition.vue'
+import OverviewRightPanelItem from '@/views/Overview/components/OverviewRightPanel/OverviewRightPanelItem.vue'
 
 const { clusterId, kubernetesUpgradeStatus, talosUpgradeStatus } = defineProps<{
   clusterId: string
